Add maven-compiler-plugin
The default source and default target settings for the Maven
Compiler Plugin are both 1.5. Explicitly declaring the plugin in the
build section of the POM allows us to change the source
and target settings to 1.8, which prepares the project to be able
to use and build Java 8 features, such as lambdas.